# Speed up IPv4Address.ipv6_mapped (numbers instead of strings)