@jstasiak 2024-05-30
I think this is small enough to not need an issue number. Since this is a 3.13+ feature I don't believe a news entry is needed either.